LF-4765 Move types and data formatting to service
