+++ packages/api/src/util/ensembleService.types.ts
@@ -120,11 +120,3 @@ export function isExternalIrrigationPrescriptionArray(
     })
   );
 }
-
-export interface AddonFunctions {
-  getIrrigationPrescriptions: (
-    farmId: string,
-    startTime?: string,
-    endTime?: string,
-  ) => Promise<IrrigationPrescription[]>;
-}
